Token refactor ThemeManager
Might not make sense, since all uses also import other things from apputils.
